% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{BaselearnerCategorical}
\alias{BaselearnerCategorical}
\title{Base-learner factory to do fit categorical features}
\format{\code{\link{S4}} object.}
\description{
\code{BaselearnerCategorical} takes a categorical feature and fits it using a L1 penalty.
}
\section{Usage}{

\preformatted{
BaselearnerCategorical$new(data_source, data_target, list(learning_rate, penalty, iters))
}
}

\section{Arguments}{

\describe{
\item{\code{data_source} [\code{Data} Object]}{
  Data object which contains the source data.
}
\item{\code{data_target} [\code{Data} Object]}{
  Data object which gets the transformed source data.
}
\item{\code{learning_rate} [\code{numeric(1)}]}{
  Learning rate of coordinate descent.
}
\item{\code{penalty} [\code{numeric(1)}]}{
  L1 penalty for the fitting process.
}
\item{\code{iters} [\code{integer(1)}]}{
  Number of iterations of coordinate descent.
}
}
}

\section{Fields}{

  This class doesn't contain public fields.
}

\section{Methods}{

\describe{
\item{\code{getData()}}{Get the data matrix of the target data which is used
  for modeling.}
\item{\code{transformData(X)}}{Transform a data matrix as defined within the
  factory. The argument has to be a matrix with one column.}
\item{\code{summarizeFactory()}}{Summarize the base-learner factory object.}
}
}

