% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{BaselearnerTensor}
\alias{BaselearnerTensor}
\title{Base learner factory to make regression using tensor products}
\format{
\code{\link{S4}} object.
}
\description{
\code{BaselearnerTensor} creates a combined base learner factory
object which can be registered within a base learner list and then used
for training.
}
