% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{CategoricalData}
\alias{CategoricalData}
\title{Data class for character variables}
\format{
\code{\link{S4}} object.
}
\description{
\code{CategoricalData} creates an data object which can be used as source
object to instantiate categorical base learner.
}
\section{Usage}{

\preformatted{
CategoricalData$new(x, data_identifier)
}
}

\section{Arguments}{

\describe{
\item{\code{x} [\code{character}]}{
  Character vector containing the classes.
}
\item{\code{data_identifier} [\code{character(1)}]}{
  The name for the data specified in \code{data_mat}. Note that it is
  important to have the same data names for train and evaluation data.
}
}
}

\section{Fields}{

  This class doesn't contain public fields.
}

\section{Methods}{

\describe{
\item{\code{getData()}}{Get numerical representation of the data.}
\item{\code{getIdentifier()}}{Get data identifier.}
\item{\code{getDictionary()}}{Get internal encoding to map numerical encoding to raw data.}
}
}

\examples{
# Sample data:
x = sample(c("one","two", "three"), 20, TRUE)

# Create new data object:
data_obj = CategoricalData$new(x, "cat")

# Get data and identifier:
data_obj$getData()
data_obj$getIdentifier()
data_obj$getDictionary()

}
