% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{CategoricalDataRaw}
\alias{CategoricalDataRaw}
\title{Data class for categorical variables

\link{CategoricalDataRaw} creates an data object which can be used as source
object to instantiate categorical base learner. In contrast to \link{CategoricalData}
the data are stored as raw categorical vector. \link{CategoricalData} can be compared
to a factor class with an integer encoding and a hash map for reassignment.}
\format{
\link{S4} object.
}
\arguments{
\item{x}{(\code{character()})\cr
Categorical vector.}
}
\description{
Data class for categorical variables

\link{CategoricalDataRaw} creates an data object which can be used as source
object to instantiate categorical base learner. In contrast to \link{CategoricalData}
the data are stored as raw categorical vector. \link{CategoricalData} can be compared
to a factor class with an integer encoding and a hash map for reassignment.
}
\section{Usage}{

\preformatted{
CategoricalDataRaw$new(x, data_identifier)
}
}

\section{Fields}{

This class doesn't contain public fields.
}

\section{Methods}{

\describe{
\itemize{
\item \verb{$getData()}: \verb{() -> stop()}\cr
Throws error because no representation is calculated.}
\item \verb{$getRawData()}: \verb{() -> character()}
\item \verb{$getIdentifier()}: \verb{() -> character(1)}
}
}

\section{Inherited methods from Data}{

\itemize{
\item \verb{$getDataType()}: \verb{() -> character(1)}
}
}

\examples{
# Sample data:
x = sample(c("one","two", "three"), 20, TRUE)

# Create new data object:
data_obj = CategoricalDataRaw$new(x, "cat_raw")

# Get data and identifier:
data_obj$getRawData()
data_obj$getIdentifier()

}
