% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{clearRegisteredFactorys}
\alias{clearRegisteredFactorys}
\title{Clear the existing Hash Map of 'BaselearnerFactorys'}
\usage{
clearRegisteredFactorys()
}
\description{
This function delets all registered factorys. This may be
  necessary for fitting a new object.
}
