% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compboost.R
\name{compboost}
\alias{compboost}
\title{Compboost main R Function}
\usage{
compboost(name)
}
\arguments{
\item{name}{[\code{character(1)}] \cr
Name to initialize the C++Class object.}
}
\value{
[\code{C++Class}] \cr
  A object of class 'CompboostWrapper'.
}
\description{
This function provides an R API to create the compboost class 
  object automatically. Notice that it is also possible to initialize 
  everything by hand'.
}
