% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{CustomCppBlearner}
\alias{CustomCppBlearner}
\title{Create custom cpp base-learner by using cpp functions and external pointer.}
\format{\code{\link{S4}} object.}
\description{
\code{CustomCppBlearner} creates a custom base-learner by using 
\code{Rcpp::XPtr} to set \code{C++} functions.
}
\section{Usage}{

\preformatted{
CustomCppBlearner(data_source, data_target, instantiate_data_ptr, train_ptr,
  predict_ptr)
}
}

\section{Arguments}{

\describe{
\item{\code{data_source} [\code{Data} Object]}{ 
  Data object which contains the source data.
}
\item{\code{data_target} [\code{Data} Object]}{ 
  Data object which gets the transformed source data.
}
\item{\code{instantiate_data_ptr} [\code{externalptr}]}{
  External pointer to the \code{C++} instantiate data function.
}
\item{\code{train_ptr} [\code{externalptr}]}{
  External pointer to the \code{C++} train function.
}
\item{\code{predict_ptr} [\code{externalptr}]}{
  External pointer to the \code{C++} predict function.
}
}
}

\section{Details}{

  For an example see the extending compboost vignette or the function
  \code{getCustomCppExample}.
  
  This class is a wrapper around the pure \code{C++} implementation. To see
  the functionality of the \code{C++} class visit
  \url{https://schalkdaniel.github.io/compboost/cpp_man/html/classblearner_1_1_custom_cpp_blearner.html}.
}

\section{Fields}{

  This class doesn't contain public fields.
}

\section{Methods}{

\describe{
\item{\code{train(response)}}{Predict parameters of the base-learner using 
  a given \code{response}.}
\item{\code{getParameter()}}{Get the estimated parameters.}
\item{\code{predict()}}{Predict by using the train data.}
\item{\code{predictNewdata(newdata)}}{Predict by using a new \code{Data}  
  object.}
\item{\code{getData()}}{Get the data matrix of the target data which is used
  for modelling.}
}
}

\examples{
# Sample data:
data.mat = cbind(1, 1:10)
y = 2 + 3 * 1:10

# Create new data object:
data.source = InMemoryData$new(data.mat, "my.data.name")
data.target = InMemoryData$new()

# Source the external pointer exposed by using XPtr:
Rcpp::sourceCpp(code = getCustomCppExample(silent = TRUE))

# Create new linear base-learner:
bl.custom.cpp = CustomCppBlearner$new(data.source, data.target, dataFunSetter(),
  trainFunSetter(), predictFunSetter())

# Train the learner:
bl.custom.cpp$train(y)

# Get estimated parameter:
bl.custom.cpp$getParameter()

}
