% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getBestBaselearner}
\alias{getBestBaselearner}
\title{Optimize over registered Baselearner}
\usage{
getBestBaselearner(pseudo_residuals)
}
\arguments{
\item{pseudo_residuals}{[\code{numeric}] \cr
Pseudo residuals to train the baselearner.}
}
\value{
[\code{list}] \cr
  List of specifica about the best baselearner.
}
\description{
This function train every baselearner within the registered
  list and returns a list of specifica about the best baselearner since
  we are not supposed to give back a not exported C++ class in R.
}
