% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{InMemoryData}
\alias{InMemoryData}
\title{Store data in RAM}
\format{
\link{S4} object.
}
\arguments{
\item{data_mat}{(\code{matrix()})\cr
The data matrix.}

\item{data_identifier}{(\code{character(1)})\cr
Data id, e.g. a feature name.}
}
\description{
This data container stores a vector as it is in the RAM and makes it
accessible for \link{Compboost}.
}
\section{Usage}{

\preformatted{
InMemoryData$new()
InMemoryData$new(data_mat, data_identifier)
InMemoryData$new(data_mat, data_identifier, use_sparse)
}
}

\section{Fields}{

This class doesn't contain public fields.
}

\section{Methods}{

\itemize{
\item \verb{$getData()}: \verb{() -> matrix()}
\item \verb{$getIdentifier()}: \verb{() -> character(1)}
}
}

\section{Inherited methods from Data}{

\itemize{
\item \verb{$getDataType()}: \verb{() -> character(1)}
}
}

\examples{
# Sample data:
data_mat = cbind(rnorm(10))

# Create new data object:
data_obj = InMemoryData$new(data_mat, "my_data_name")

# Get data and identifier:
data_obj$getData()
data_obj$getIdentifier()

}
