% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlr_learner.R
\name{LearnerClassifCompboost}
\alias{LearnerClassifCompboost}
\title{Component-wise gradient boosting classification learner}
\description{
A \link[mlr3:Learner]{Learner} for a component-wise boosting model implemented in \link{Compboost}.
}
\examples{
task = mlr3::tsk("german_credit")
l = mlr3::lrn("classif.compboost", baselearner = "components",
  df = 5, iterations = 20)
l$train(task)
l$selected_features()
}
\section{Super classes}{
\code{\link[mlr3:Learner]{mlr3::Learner}} -> \code{\link[compboost:LearnerCompboost]{compboost::LearnerCompboost}} -> \code{LearnerClassifCompboost}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-LearnerClassifCompboost-new}{\code{LearnerClassifCompboost$new()}}
\item \href{#method-LearnerClassifCompboost-clone}{\code{LearnerClassifCompboost$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="base_learner"><a href='../../mlr3/html/Learner.html#method-Learner-base_learner'><code>mlr3::Learner$base_learner()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="format"><a href='../../mlr3/html/Learner.html#method-Learner-format'><code>mlr3::Learner$format()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="help"><a href='../../mlr3/html/Learner.html#method-Learner-help'><code>mlr3::Learner$help()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="predict"><a href='../../mlr3/html/Learner.html#method-Learner-predict'><code>mlr3::Learner$predict()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="predict_newdata"><a href='../../mlr3/html/Learner.html#method-Learner-predict_newdata'><code>mlr3::Learner$predict_newdata()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="print"><a href='../../mlr3/html/Learner.html#method-Learner-print'><code>mlr3::Learner$print()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="reset"><a href='../../mlr3/html/Learner.html#method-Learner-reset'><code>mlr3::Learner$reset()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="train"><a href='../../mlr3/html/Learner.html#method-Learner-train'><code>mlr3::Learner$train()</code></a></span></li>
<li><span class="pkg-link" data-pkg="compboost" data-topic="LearnerCompboost" data-id="importance"><a href='../../compboost/html/LearnerCompboost.html#method-LearnerCompboost-importance'><code>compboost::LearnerCompboost$importance()</code></a></span></li>
<li><span class="pkg-link" data-pkg="compboost" data-topic="LearnerCompboost" data-id="save_json"><a href='../../compboost/html/LearnerCompboost.html#method-LearnerCompboost-save_json'><code>compboost::LearnerCompboost$save_json()</code></a></span></li>
<li><span class="pkg-link" data-pkg="compboost" data-topic="LearnerCompboost" data-id="selected_features"><a href='../../compboost/html/LearnerCompboost.html#method-LearnerCompboost-selected_features'><code>compboost::LearnerCompboost$selected_features()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LearnerClassifCompboost-new"></a>}}
\if{latex}{\out{\hypertarget{method-LearnerClassifCompboost-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of this \link[R6:R6Class]{R6} class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LearnerClassifCompboost$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LearnerClassifCompboost-clone"></a>}}
\if{latex}{\out{\hypertarget{method-LearnerClassifCompboost-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LearnerClassifCompboost$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
