% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlr_learner.R
\name{LearnerCompboost}
\alias{LearnerCompboost}
\title{Component-wise gradient boosting learner}
\description{
A \link[mlr3:Learner]{Learner} for a component-wise boosting model implemented in \link{Compboost}.
}
\section{Super class}{
\code{\link[mlr3:Learner]{mlr3::Learner}} -> \code{LearnerCompboost}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-LearnerCompboost-new}{\code{LearnerCompboost$new()}}
\item \href{#method-LearnerCompboost-importance}{\code{LearnerCompboost$importance()}}
\item \href{#method-LearnerCompboost-selected_features}{\code{LearnerCompboost$selected_features()}}
\item \href{#method-LearnerCompboost-save_json}{\code{LearnerCompboost$save_json()}}
\item \href{#method-LearnerCompboost-clone}{\code{LearnerCompboost$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="base_learner"><a href='../../mlr3/html/Learner.html#method-Learner-base_learner'><code>mlr3::Learner$base_learner()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="format"><a href='../../mlr3/html/Learner.html#method-Learner-format'><code>mlr3::Learner$format()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="help"><a href='../../mlr3/html/Learner.html#method-Learner-help'><code>mlr3::Learner$help()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="predict"><a href='../../mlr3/html/Learner.html#method-Learner-predict'><code>mlr3::Learner$predict()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="predict_newdata"><a href='../../mlr3/html/Learner.html#method-Learner-predict_newdata'><code>mlr3::Learner$predict_newdata()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="print"><a href='../../mlr3/html/Learner.html#method-Learner-print'><code>mlr3::Learner$print()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="reset"><a href='../../mlr3/html/Learner.html#method-Learner-reset'><code>mlr3::Learner$reset()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="train"><a href='../../mlr3/html/Learner.html#method-Learner-train'><code>mlr3::Learner$train()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LearnerCompboost-new"></a>}}
\if{latex}{\out{\hypertarget{method-LearnerCompboost-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of this \link[R6:R6Class]{R6} class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LearnerCompboost$new(ttype)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ttype}}{(\code{character(1)}) \cr
The task type of the learner. Must be \code{regr} or \code{classif}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LearnerCompboost-importance"></a>}}
\if{latex}{\out{\hypertarget{method-LearnerCompboost-importance}{}}}
\subsection{Method \code{importance()}}{
The importance scores are extracted from the model slot \verb{$calculateFEatureImportance()}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LearnerCompboost$importance()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Named \code{numeric()}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LearnerCompboost-selected_features"></a>}}
\if{latex}{\out{\hypertarget{method-LearnerCompboost-selected_features}{}}}
\subsection{Method \code{selected_features()}}{
Selected features are extracted from the model slot \verb{$getSelectedBaselearner()}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LearnerCompboost$selected_features()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{character()}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LearnerCompboost-save_json"></a>}}
\if{latex}{\out{\hypertarget{method-LearnerCompboost-save_json}{}}}
\subsection{Method \code{save_json()}}{
Save the model to a JSON file.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LearnerCompboost$save_json(file)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{file}}{(\code{character(1)}\cr
Name/path to the file.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{character(1)}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LearnerCompboost-clone"></a>}}
\if{latex}{\out{\hypertarget{method-LearnerCompboost-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LearnerCompboost$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
