% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_from_cpp.R
\name{loadFromJson}
\alias{loadFromJson}
\title{Load Compboost model from JSON}
\usage{
loadFromJson(file)
}
\arguments{
\item{file}{(`character(1)`)\cr
A data frame containing the data.}
}
\value{
A model of the \code{Compboost} class. This model is an \code{R6} object
  which can be used for retraining, predicting, plotting, and anything described in
  \code{?Compboost}.
}
\description{
Load a [Compboost] object from a JSON file. Because of the underlying \code{C++} objects,
it is not possible to use \code{R}'s native load and save methods.
}
\examples{
cboost = boostLinear(data = iris, target = "Sepal.Length", loss = LossQuadratic$new(),
  oob_fraction = 0.3)
cboost$model$saveJson("cboost.json")
cboost2 = loadFromJson("cboost.json")
file.remove("cboost.json")
}
