% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{LoggerIteration}
\alias{LoggerIteration}
\title{Logger class to log the current iteration}
\format{
\link{S4} object.
}
\arguments{
\item{logger_id}{(\code{character(1)})\cr
Identifier of the logger.}

\item{use_as_stopper}{(\code{logical(1)})\cr
Boolean to indicate if the logger should also be used as a stopper.}

\item{max_iterations}{(\code{integer(1)})\cr
If the logger is used as stopper this argument defines the maximal iterations.}
}
\description{
Logger class to log the current iteration
}
\section{Usage}{

\preformatted{
LoggerIterationWrapper$new(logger_id, use_as_stopper, max_iterations)
}
}

\section{Fields}{

This class doesn't contain public fields.
}

\section{Methods}{

\itemize{
\item \verb{$summarizeLogger()}: \verb{() -> ()}
}
}

\examples{
# Define logger:
log_iters = LoggerIteration$new("iterations", FALSE, 100)

# Summarize logger:
log_iters$summarizeLogger()

}
