% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{LoggerList}
\alias{LoggerList}
\title{Collect loggers}
\format{
\link{S4} object.
}
\description{
This class collects all loggers that are used in the algorithm and
takes care about stopping strategies and tracing.
}
\section{Usage}{

\preformatted{
LoggerList$new()
}
}

\section{Fields}{

This class doesn't contain public fields.
}

\section{Methods}{

\itemize{
\item \verb{$registerLogger()}: \verb{Logger* -> ()}
\item \verb{$printRegisteredLogger()}: \verb{() -> ()}
\item \verb{$clearRegisteredLogger()}: \verb{() -> ()}
\item \verb{$getNumberOfRegisteredLogger()}: \verb{() -> integer(1)}
\item \verb{$getNamesOfRegisteredLogger()}: \verb{() -> character()}
\item \verb{$isStopper()}: \verb{() -> logical()}
}
}

\examples{
# Define logger:
log_iters = LoggerIteration$new("iteration", TRUE, 100)
log_time = LoggerTime$new("time", FALSE, 20, "minutes")

# Create logger list:
logger_list = LoggerList$new()

# Register new loggeR:
logger_list$registerLogger(log_iters)
logger_list$registerLogger(log_time)

# Print registered logger:
logger_list$printRegisteredLogger()

# Remove all logger:
logger_list$clearRegisteredLogger()

# Get number of registered logger:
logger_list$getNumberOfRegisteredLogger()

}
