% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{LoggerTime}
\alias{LoggerTime}
\title{Log the runtime}
\format{
\link{S4} object.
}
\arguments{
\item{logger_id}{(\code{character(1)})\cr
Identifier of the logger.}

\item{param-use_as_stopper}{(\code{logical(1)})\cr
Boolean to indicate if the logger should also be used as a stopper.}

\item{max_time}{(\code{integer(1)})\cr
If the logger is used as stopper this argument contains the maximal time
which are available to train the model.}

\item{time_unit}{(\code{character(1)})\cr
The unit in which the time is measured. Choices are \code{minutes},
\code{seconds} or \code{microseconds}.}
}
\description{
This class logs the runtime of the algorithm. The logger also can be used
to stop the algorithm after a defined time budget. The available time units are:
\itemize{
\item minutes
\item seconds
\item microseconds
}
}
\section{Usage}{

\preformatted{
LoggerTime$new(logger_id, use_as_stopper, max_time, time_unit)
}
}

\section{Fields}{

This class doesn't contain public fields.
}

\section{Methods}{

\itemize{
\item \verb{$summarizeLogger()}: \verb{() -> ()}
}
}

\examples{
# Define logger:
log_time = LoggerTime$new("time_minutes", FALSE, 20, "minutes")

# Summarize logger:
log_time$summarizeLogger()

}
