% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{LossAbsolute}
\alias{LossAbsolute}
\title{Absolute loss for regression tasks.}
\format{
\link{S4} object.
}
\arguments{
\item{offset}{(\code{numeric(1)} | \code{matrix()})\cr
Numerical value or matrix to set a custom offset. If used, this
value is returned instead of the loss optimal initialization.}
}
\description{
This loss can be used for regression with \eqn{y \in \mathrm{R}}.
}
\details{
\strong{Loss Function:}
\deqn{
  L(y, f(x)) = | y - f(x)|
}
\strong{Gradient:}
\deqn{
  \frac{\delta}{\delta f(x)}\ L(y, f(x)) = -\mathrm{sign}(y - f(x))
}
\strong{Initialization:}
\deqn{
  \hat{f}^{[0]}(x) = \mathrm{arg~min}_{c\in R}\ \frac{1}{n}\sum\limits_{i=1}^n
  L(y^{(i)}, c) = \mathrm{median}(y)
}
}
\section{Usage}{

\preformatted{
LossAbsolute$new()
LossAbsolute$new(offset)
}
}

\section{Inherited methods from Loss}{

\itemize{
\item \verb{$loss()}: \verb{matrix(), matrix() -> matrix()}
\item \verb{$gradient()}: \verb{matrix(), matrix() -> matrix()}
\item \verb{$constInit()}: \code{matrix() -> matrix()}
\item \verb{$calculatePseudoResiduals()}: \verb{matrix(), matrix() -> matrix()}
\item \verb{$getLossType()}: \verb{() -> character(1)}
}
}

\examples{

# Create new loss object:
absolute_loss = LossAbsolute$new()
absolute_loss

}
