% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{LossCustomCpp}
\alias{LossCustomCpp}
\title{Create custom cpp losses by using cpp functions and external pointer.}
\format{
\code{\link{S4}} object.
}
\description{
\code{LossCustomCpp} creates a custom loss by using
\code{Rcpp::XPtr} to set \code{C++} functions.
}
\section{Usage}{

\preformatted{
LossCustomCpp$new(loss_ptr, grad_ptr, const_init_ptr)
}
}

\section{Arguments}{

\describe{
\item{\code{loss_ptr} \link{\code{externalptr}}}{
External pointer to the \code{C++} loss function.
}
\item{\code{grad_ptr} \link{\code{externalptr}}}{
External pointer to the \code{C++} gradient function.
}
\item{\code{const_init_ptr} \link{\code{externalptr}}}{
External pointer to the \code{C++} constant initialization function.
}
}
}

\examples{
\dontrun{
# Load loss functions:
Rcpp::sourceCpp(code = getCustomCppExample(example = "loss", silent = TRUE))

# Create new custom quadratic loss:
my_cpp_loss = LossCustomCpp$new(lossFunSetter(), gradFunSetter(), constInitFunSetter())
}
}
