% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{LossQuantile}
\alias{LossQuantile}
\title{Quantile loss for regression tasks.}
\format{\code{\link{S4}} object.}
\description{
This loss can be used for regression with \eqn{y \in \mathrm{R}}.
}
\details{
\strong{Loss Function:}
\deqn{
  L(y, f(x)) = h| y - f(x)|
}
\strong{Gradient:}
\deqn{
  \frac{\delta}{\delta f(x)}\ L(y, f(x)) = -h\mathrm{sign}( y - f(x))
}
\strong{Initialization:}
\deqn{
  \hat{f}^{[0]}(x) = \mathrm{arg~min}_{c\in R}\ \frac{1}{n}\sum\limits_{i=1}^n
  L(y^{(i)}, c) = \mathrm{quantile}(y, q)
}
}
\section{Usage}{

\preformatted{
LossAbsolute$new()
LossAbsolute$new(quantile)
LossAbsolute$new(offset, quantile)
}
}

\section{Arguments}{

\describe{
\item{\code{offset} [\code{numeric(1)}]}{
  Numerical value which can be used to set a custom offset. If so, this
  value is returned instead of the loss optimal initialization.
}
\item{\code{quantile} [\code{numeric(1)}]}{
  Numerical value between 0 and 1 indicating the quantile used for boosting.
}
}
}

\examples{

# Create new loss object:
quadratic_loss = LossQuadratic$new()
quadratic_loss

}
