% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlr_learner.R
\name{mlr_learners_regr.compboost}
\alias{mlr_learners_regr.compboost}
\alias{LearnerRegrCompboost}
\title{Regression component-wise boosting learner}
\description{
A \link{Learner} for a component-wise boosting model implemented in \link{Compboost}
in package \CRANpkg{compboost}.
}
\examples{
l = lrn("regr.compboost", baselearner = "linear", iterations = 50)
task = tsk("mtcars")
l$train(task)
l$importance()
}
\section{Super classes}{
\code{\link[mlr3:Learner]{mlr3::Learner}} -> \code{\link[compboost:LearnerCompboost]{compboost::LearnerCompboost}} -> \code{LearnerRegrCompboost}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{LearnerRegrCompboost$new()}}
\item \href{#method-clone}{\code{LearnerRegrCompboost$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="base_learner">}\href{../../mlr3/html/Learner.html#method-base_learner}{\code{mlr3::Learner$base_learner()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="format">}\href{../../mlr3/html/Learner.html#method-format}{\code{mlr3::Learner$format()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="help">}\href{../../mlr3/html/Learner.html#method-help}{\code{mlr3::Learner$help()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="predict">}\href{../../mlr3/html/Learner.html#method-predict}{\code{mlr3::Learner$predict()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="predict_newdata">}\href{../../mlr3/html/Learner.html#method-predict_newdata}{\code{mlr3::Learner$predict_newdata()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="print">}\href{../../mlr3/html/Learner.html#method-print}{\code{mlr3::Learner$print()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="reset">}\href{../../mlr3/html/Learner.html#method-reset}{\code{mlr3::Learner$reset()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="train">}\href{../../mlr3/html/Learner.html#method-train}{\code{mlr3::Learner$train()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="compboost" data-topic="LearnerCompboost" data-id="importance">}\href{../../compboost/html/LearnerCompboost.html#method-importance}{\code{compboost::LearnerCompboost$importance()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="compboost" data-topic="LearnerCompboost" data-id="save_json">}\href{../../compboost/html/LearnerCompboost.html#method-save_json}{\code{compboost::LearnerCompboost$save_json()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="compboost" data-topic="LearnerCompboost" data-id="selected_features">}\href{../../compboost/html/LearnerCompboost.html#method-selected_features}{\code{compboost::LearnerCompboost$selected_features()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of this \link[R6:R6Class]{R6} class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LearnerRegrCompboost$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LearnerRegrCompboost$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
