% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{OptimizerAGBM}
\alias{OptimizerAGBM}
\title{Nesterovs momentum}
\format{
\link{S4} object.
}
\arguments{
\item{ncores}{(\code{integer(1)})\cr
Number of cores over which the base learner fitting is distributed.}

\item{momentum}{(\code{numeric(1)})\cr
Momentum term used to accelerate the fitting process. If chosen large, the algorithm trains
faster but also tends to overfit faster.}
}
\description{
This class defines a new object to conduct Nesterovs momentum in function space.
}
\section{Usage}{

\preformatted{
OptimizerAGBM$new(momentum)
OptimizerAGBM$new(momentum, ncores)
}
}

\section{Fields}{

This class doesn't contain public fields.
}

\section{Methods}{

\itemize{
\item \verb{$getOptimizerType()}: \verb{() -> character(1)}
\item \verb{$getStepSize()}: \verb{() -> numeric()}
\item \verb{$getMomentumParameter()}: \verb{() -> numeric(1)}
\item \verb{$getSelectedMomentumBaselearner()}: \verb{() -> character()}
\item \verb{$getParameterMatrix()}: \verb{() -> list(matrix()}
\item \verb{$getErrorCorrectedPseudoResiduals()}: \verb{() -> matrix()}
}
}

\examples{

optimizer = OptimizerAGBM$new(0.1)

}
