% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{OptimizerCoordinateDescent}
\alias{OptimizerCoordinateDescent}
\title{Coordinate descent}
\format{
\link{S4} object.
}
\arguments{
\item{ncores}{(\code{integer(1)})\cr
Number of cores over which the base learner fitting is distributed.}
}
\description{
This class defines a new object to conduct gradient descent in function space.
Because of the component-wise structure, this is more like a block-wise
coordinate descent.
}
\section{Usage}{

\preformatted{
OptimizerCoordinateDescent$new()
OptimizerCoordinateDescent$new(ncores)
}
}

\section{Fields}{

This class doesn't contain public fields.
}

\section{Methods}{

\itemize{
\item \verb{$getOptimizerType()}: \verb{() -> character(1)}
\item \verb{$getStepSize()}: \verb{() -> numeric()}
}
}

\examples{

# Define optimizer:
optimizer = OptimizerCoordinateDescent$new()

}
