% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{OptimizerCoordinateDescentLineSearch}
\alias{OptimizerCoordinateDescentLineSearch}
\title{Coordinate descent with line search}
\format{
\link{S4} object.
}
\arguments{
\item{ncores}{(\code{integer(1)})\cr
Number of cores over which the base learner fitting is distributed.}
}
\description{
Same as \link{OptimizerCoordinateDescent} but with a line search in each iteration.
}
\section{Usage}{

\preformatted{
OptimizerCoordinateDescentLineSearch$new()
OptimizerCoordinateDescentLineSearch$new(ncores)
}
}

\section{Fields}{

This class doesn't contain public fields.
}

\section{Methods}{

\itemize{
\item \verb{$getOptimizerType()}: \verb{() -> character(1)}
\item \verb{$getStepSize()}: \verb{() -> numeric()}
}
}

\examples{

# Define optimizer:
optimizer = OptimizerCoordinateDescentLineSearch$new()

}
