% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{playArma}
\alias{playArma}
\title{RcppArmadillo test}
\usage{
playArma(x)
}
\arguments{
\item{x}{[\code{numeric}] \cr
Numeric vector for the armadillo to play around with.}
}
\value{
[\code{numeric}] \cr
  The vector after the torture.
}
\description{
This function is used to play around with Armadillo.
}
