% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pe_uni.R
\name{plotBaselearner}
\alias{plotBaselearner}
\title{Visualize contribution of one base learner}
\usage{
plotBaselearner(cboost, blname, npoints = 100L)
}
\arguments{
\item{cboost}{(\link{Compboost})\cr
A trained \link{Compboost} object.}

\item{blname}{(\code{character(1L)})\cr
Name of the base learner. Must be one of \code{cboost$getBaselearnerNames()}.}

\item{npoints}{(\code{integer(1L)})\cr
Number of points which are predicted for the lines (only applies to numerical features).}
}
\value{
\code{ggplot} object containing the graphic.
}
\description{
This function visualizes the contribution of a base learner to the overall
prediction score. For visualization of partial effects see \code{\link[=plotPEUni]{plotPEUni()}}.
}
\examples{
cboost = Compboost$new(data = iris, target = "Petal.Length",
  loss = LossQuadratic$new())
cboost$addComponents("Sepal.Width")
cboost$train(500L)
plotBaselearner(cboost, "Sepal.Width_linear")
plotBaselearner(cboost, "Sepal.Width_Sepal.Width_spline_centered")
}
