% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_bltraces.R
\name{plotBaselearnerTraces}
\alias{plotBaselearnerTraces}
\title{Visualize base learner traces}
\usage{
plotBaselearnerTraces(cboost, value = 1, n_legend = 5L)
}
\arguments{
\item{cboost}{[\code{Compboost} class]\cr
A trained \code{Compboost} object.}

\item{n_legend}{[\code{integer(1L)}]\cr
Number of colored base learners added to the legend.}

\item{values}{[\code{numeric(1L) | numeric(length(cboost$getSelectedBaselearner()))}]\cr
Value used to show the base learner development w.r.t. to the value.}
}
\value{
\code{ggplot} object containing the graphic.
}
\description{
This function shows how the base learners evolves over the fitting process.
The default is to show how the frequency of a single base learner included
into the model evolves. Additionally, with the \code{value} argument, vectors
(e.g. the risk) can be used to show how the base learner specific risk reduction
evolves during the fitting process.
}
\examples{
cboost = Compboost$new(data = iris, target = "Petal.Length",
  loss = LossQuadratic$new())
cboost$addComponents("Sepal.Width")
cboost$addBaselearner("Species", "ridge", BaselearnerCategoricalRidge)
cboost$train(500L)
plotBaselearnerTraces(cboost)
}
