% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_plotter.R
\name{plotCompboostParameter}
\alias{plotCompboostParameter}
\title{Parameter plotter for a trained compboost object}
\usage{
plotCompboostParameter(object, legend = TRUE, ...)
}
\arguments{
\item{object}{[\code{character(1)}] \cr
Trained compboost object.}

\item{legend}{[\code{logical(1)}] \cr
Logical to specify if a legend should be plotted.}

\item{...}{Additional parameter given to plot.}
}
\description{
This function can be used to print the trace of the parameters
  of a trained compboost object.
}
