% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pe_vi.R
\name{plotFeatureImportance}
\alias{plotFeatureImportance}
\title{Visualize the feature importance}
\usage{
plotFeatureImportance(cboost, num_feats = NULL, aggregate = TRUE)
}
\arguments{
\item{cboost}{[\code{Compboost} class]\cr
A trained \code{Compboost} object.}

\item{num_feats}{[\code{integer(1L)}]\cr
Number of features that are visualized. All features are added if set to \code{NULL}.}

\item{aggregate}{[\code{logical(1L)}]\cr
Flag whether the feature importance is aggregated by feature. Otherwise it is
visualized per base learner.}
}
\value{
\code{ggplot} object containing the graphic.
}
\description{
This function visualizes the feature importance as horizontal bar plot.
}
\examples{
cboost = boostSplines(data = iris, target = "Sepal.Length", loss = LossQuadratic$new())
plotFeatureImportance(cboost)
plotFeatureImportance(cboost, num_feats = 2)
plotFeatureImportance(cboost, num_feats = 2, aggregate = FALSE)
}
