% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pe_uni.R
\name{plotPEUni}
\alias{plotPEUni}
\title{Visualize partial effect of a feature}
\usage{
plotPEUni(cboost, feat, npoints = 100L, individual = TRUE)
}
\arguments{
\item{cboost}{\link{\code{Compboost} class}\cr
A trained \code{Compboost} object.}

\item{feat}{\link{\code{character(1L)}}\cr
Name of the feature.}

\item{npoints}{\link{\code{integer(1L)}}\cr
Number of points which are predicted for the lines (only applies to numerical features).}

\item{individual}{\link{\code{logical(1L)}}\cr
Flag whether individual base learners should be added to the graphic or not.}
}
\value{
\code{ggplot} object containing the graphic.
}
\description{
This function visualizes the contribution of a specific feature to the overall
prediction score. If multiple base learner of the same features are included,
they are all added to the graphic as well as the aggregated contribution. The
difference to \code{plotBaselearner} is that potentially multiple base learners
that are based on \code{feat} are aggregated and visualized while \code{plotBaselearner}
only visualizes the contribution of one specific base learner. The function
also automatically decides whether the given feature is numeric or categorical and
chooses an appropriate technique (lines for numeric and horizontal lines for categorical).
}
\examples{
cboost = Compboost$new(data = iris, target = "Petal.Length",
  loss = LossQuadratic$new())
cboost$addComponents("Sepal.Width")
cboost$train(500L)
plotPEUni(cboost, "Sepal.Width")
}
