% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_risk.R
\name{plotRisk}
\alias{plotRisk}
\title{Visualize the risk}
\usage{
plotRisk(cboost)
}
\arguments{
\item{cboost}{[\code{Compboost} class]\cr
A trained \code{Compboost} object.}
}
\value{
\code{ggplot} object containing the graphic.
}
\description{
This function visualizes the risk during training. If validation data are given, then
the train risk is plotted against the validation risk.
}
\examples{
cboost_no_valdat = boostSplines(data = iris, target = "Sepal.Length",
  loss = LossQuadratic$new())
plotRisk(cboost_no_valdat)

cboost_valdat = boostSplines(data = iris, target = "Sepal.Length",
  loss = LossQuadratic$new(), oob_fraction = 0.3)
plotRisk(cboost_valdat)
}
