% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{printRegisteredFactorys}
\alias{printRegisteredFactorys}
\title{Print Registered Factorys}
\usage{
printRegisteredFactorys()
}
\description{
This function prints all registered factorys to the screen.
  This is just ment to be a control function to see which factorys are
  already initialized.
}
