% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{QuadraticLoss}
\alias{QuadraticLoss}
\title{Quadratic loss for regression tasks.}
\format{\code{\link{S4}} object.}
\description{
This loss can be used for regression with \eqn{y \in \mathrm{R}}.
}
\details{
\strong{Loss Function:}
\deqn{
  L(y, f(x)) = \frac{1}{2}( y - f(x))^2
}
\strong{Gradient:}
\deqn{
  \frac{\delta}{\delta f(x)}\ L(y, f(x)) = f(x) - y
}
\strong{Initialization:}
\deqn{
  \hat{f}^{[0]}(x) = \mathrm{arg~min}{c\in\mathrm{R}}{\mathrm{arg~min}}\ \frac{1}{n}\sum\limits_{i=1}^n
  L\left(y^{(i)}, c\right) = \bar{y}
}
}
\section{Usage}{

\preformatted{
QuadraticLoss$new()
QuadraticLoss$new(offset)
}
}

\section{Arguments}{

\describe{
\item{\code{offset} [\code{numeric(1)}]}{ 
  Numerical value which can be used to set a custom offset. If so, this 
  value is returned instead of the loss optimal initialization.
}
}
}

\section{Details}{

     
  This class is a wrapper around the pure \code{C++} implementation. To see
  the functionality of the \code{C++} class visit
  \url{https://schalkdaniel.github.io/compboost/cpp_man/html/classloss_1_1_quadratic_loss.html}.
}

\section{Fields}{

  This class doesn't contain public fields.
}

\section{Methods}{

\describe{
\item{\code{testLoss(truth, prediction)}}{Calculates the loss for a given
  true response and a corresponding prediction.}
\item{\code{testGradient(truth, prediction)}}{Calculates the gradient of 
  loss function for a given true response and a corresponding prediction.}
\item{\code{testconstantInitializer(truth)}}{Calculates the constant 
  initialization of a given vector of true values.}
}
}

\examples{
# Sample data:
set.seed(123)
truth = 1:10
prediction = truth - rnorm(10)

# Create new loss object:
quadratic.loss = QuadraticLoss$new()
quadratic.loss

# Calculate loss:
quadratic.loss$testLoss(truth, prediction)

# Calculate gradient:
quadratic.loss$testGradient(truth, prediction)

# Calculate the offset if this loss is used for the training:
quadratic.loss$testConstantInitializer(truth)

}
