% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ResponseBinaryClassif}
\alias{ResponseBinaryClassif}
\title{Create response object for binary classification.}
\format{
\link{S4} object.
}
\description{
\code{ResponseBinaryClassif} creates a response object that are used as target during the
fitting process.
}
\section{Usage}{

\preformatted{
ResponseBinaryClassif$new(target_name, pos_class, response)
ResponseBinaryClassif$new(target_name, pos_class, response, weights)
}
}

\examples{

response_binary = ResponseBinaryClassif$new("target", "A", sample(c("A", "B"), 10, TRUE))
response_binary$getResponse()
response_binary$getPrediction()
response_binary$getPredictionTransform() # Applies sigmoid to prediction scores
response_binary$getPredictionResponse()  # Categorizes depending on the transformed predictions
response_binary$getTargetName()
response_binary$setThreshold(0.7)
response_binary$getThreshold()
response_binary$getPositiveClass()

}
