% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{timesTwo}
\alias{timesTwo}
\title{timesTwo test}
\usage{
timesTwo(x)
}
\arguments{
\item{x}{[\code{numeric}] \cr
Vector which should be multiplied by 2.}
}
\value{
New vector.
}
\description{
Multiplied by two!
}
