% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{timesTwoPtr}
\alias{timesTwoPtr}
\title{timesTwo test}
\usage{
timesTwoPtr(x)
}
\arguments{
\item{x}{[\code{numeric}] \cr
Vector which should be multiplied by 2.}
}
\value{
[\code{numeric}] \cr
  New vector.
}
\description{
Multiplied by two!
}
