% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{BaselearnerCategoricalRidge}
\alias{BaselearnerCategoricalRidge}
\title{Base-learner factory for categorical feature using Ridge penalty}
\format{
\code{\link{S4}} object.
}
\description{
\code{BaselearnerCategoricalRidge} can be used to estimate effects of  categorical
features. The categories are included as in the linear model by using a binary matrix.
The Ridge penalty enables unbiased feature selection by setting the penalty corresponding
to degree of freedoms.
}
\section{Usage}{

\preformatted{
BaselearnerCategoricalRidge$new(data_source, list(df))
}
}

\section{arguments}{

\describe{
\item{\code{data_source} [\code{data} object]}{
  data object which contains the source data.
}
}
}

\section{Fields}{

  This class doesn't contain public fields.
}

\section{Methods}{

\describe{
\item{\code{getData()}}{Get the data matrix of the target data which is used
  for modeling.}
\item{\code{transformData(X)}}{This class does is not allowed to transform data. This is due to the internal structure.}
\item{\code{summarizeFactory()}}{Summarize the base-learner factory object.}
}
}

\examples{
x = sample(c("one","two"), 20, TRUE)
ds = CategoricalData$new(x, "cat")
bl = BaselearnerCategoricalRidge$new(ds, list(df = 1))

bl$getData()
bl$summarizeFactory()

}
