% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{CategoricalDataRaw}
\alias{CategoricalDataRaw}
\title{Data class for categorical variables}
\format{
\link{S4} object.
}
\arguments{
\item{x}{(\code{character()})\cr
Categorical vector.}

\item{data_identifier}{(\code{character(1)})\cr
Data id, e.g. a feature name.}
}
\description{
\link{CategoricalDataRaw} creates an data object which can be used as source
object to instantiate categorical base learner.
}
\section{Usage}{

\preformatted{
CategoricalDataRaw$new(x, data_identifier)
}
}

\section{Fields}{

This class doesn't contain public fields.
}

\section{Methods}{

\itemize{
\item \verb{$getData()}: \verb{() -> stop()}\cr Throws error because no representation is calculated.
\item \verb{$getRawData()}: \verb{() -> character()}
\item \verb{$getIdentifier()}: \verb{() -> character(1)}
}
}

\section{Inherited methods from Data}{

\itemize{
\item \verb{$getDataType()}: \verb{() -> character(1)}
}
}

\examples{
# Sample data:
x = sample(c("one","two", "three"), 20, TRUE)

# Create new data object:
data_obj = CategoricalDataRaw$new(x, "cat_raw")

# Get data and identifier:
data_obj$getRawData()
data_obj$getIdentifier()

}
