% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{CategoricalDataRaw}
\alias{CategoricalDataRaw}
\title{Data class for character variables}
\format{
\code{\link{S4}} object.
}
\description{
\code{CategoricalDataRaw} creates an data object which can be used as source
object to instantiate categorical base learner. In contrast to \code{CategoricalData}
the data are stored as raw categorical vector.
}
\section{Usage}{

\preformatted{
CategoricalDataRaw$new(x, data_identifier)
}
}

\section{Arguments}{

\describe{
\item{\code{x} [\code{character}]}{
  Character vector containing the classes.
}
\item{\code{data_identifier} [\code{character(1)}]}{
  The name for the data specified in \code{data_mat}. Note that it is
  important to have the same data names for train and evaluation data.
}
}
}

\section{Fields}{

  This class doesn't contain public fields.
}

\section{Methods}{

\describe{
\item{\code{getData()}}{Throws error because no representation is calculated.}
\item{\code{getRawData()}}{Get raw character data.}
\item{\code{getIdentifier()}}{Get data identifier.}
}
}

\examples{
# Sample data:
x = sample(c("one","two", "three"), 20, TRUE)

# Create new data object:
data_obj = CategoricalDataRaw$new(x, "cat_raw")

# Get data and identifier:
data_obj$getRawData()
data_obj$getIdentifier()

}
