% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_custom_cpp_example.R
\name{getCustomCppExample}
\alias{getCustomCppExample}
\title{Get example C++ script to define a custom cpp logger}
\usage{
getCustomCppExample(example = "blearner", silent = FALSE)
}
\arguments{
\item{example}{[\code{character(1)}] \cr
Character value indicating if an example for the base-learner or for the
loss should be returned. The values, for example, has to be one of 
\code{blearner} or \code{loss}.}

\item{silent}{[\code{logical(1)}] \cr
Logical value indicating if the example code should be printed to the
screen.}
}
\value{
This function returns a string which can be compiled using
  \code{Rcpp::sourceCpp(code = getCustomCppExample)} to define a new 
  custom cpp logger.
}
\description{
This function can be used to print the trace of the parameters
  of a trained compboost object.
}
