% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{LoggerIteration}
\alias{LoggerIteration}
\title{Logger class to log the current iteration}
\format{
\code{\link{S4}} object.
}
\description{
Logger class to log the current iteration
}
\section{Usage}{

\preformatted{
LoggerIterationWrapper$new(logger_id, use_as_stopper, max_iterations)
}
}

\section{Arguments}{

\describe{
\item{\code{logger_id} \link{\code{character(1)}}}{
Unique identifier of the logger.
}
\item{\code{use_as_stopper} \link{\code{logical(1)}}}{
Boolean to indicate if the logger should also be used as stopper.
}
\item{\code{max_iterations} \link{\code{integer(1)}}}{
If the logger is used as stopper this argument defines the maximal
iterations.
}
}
}

\section{Fields}{

This class doesn't contain public fields.
}

\section{Methods}{

\describe{
\item{\code{summarizeLogger()}}{Summarize the logger object.}
}
}

\examples{
# Define logger:
log_iters = LoggerIteration$new("iterations", FALSE, 100)

# Summarize logger:
log_iters$summarizeLogger()

}
