% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{LoggerList}
\alias{LoggerList}
\title{Logger list class to collect all loggers}
\format{
\code{\link{S4}} object.
}
\description{
This class is meant to define all logger which should be used to track the
progress of the algorithm.
}
\section{Usage}{

\preformatted{
LoggerList$new()
}
}

\section{Fields}{

This class doesn't contain public fields.
}

\section{Methods}{

\describe{
\item{\code{clearRegisteredLogger()}}{Removes all registered logger
from the list. The used logger are not deleted, just removed from the
map.}
\item{\code{getNamesOfRegisteredLogger()}}{Returns the registered logger
names as character vector.}
\item{\code{getNumberOfRegisteredLogger()}}{Returns the number of registered
logger as integer.}
\item{\code{printRegisteredLogger()}}{Prints all registered logger.}
\item{\code{registerLogger(logger)}}{Includes a new \code{logger}
into the logger list with the \code{logger_id} as key.}
}
}

\examples{
# Define logger:
log_iters = LoggerIteration$new("iteration", TRUE, 100)
log_time = LoggerTime$new("time", FALSE, 20, "minutes")

# Create logger list:
logger_list = LoggerList$new()

# Register new loggeR:
logger_list$registerLogger(log_iters)
logger_list$registerLogger(log_time)

# Print registered logger:
logger_list$printRegisteredLogger()

# Remove all logger:
logger_list$clearRegisteredLogger()

# Get number of registered logger:
logger_list$getNumberOfRegisteredLogger()

}
