% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{LossCustomCpp}
\alias{LossCustomCpp}
\title{Custom loss using \verb{C++} functions.}
\format{
\link{S4} object.
}
\arguments{
\item{loss_ptr}{(\code{externalptr})\cr
External pointer to the \code{C++} loss function.}

\item{grad_ptr}{(\code{externalptr})\cr
External pointer to the \code{C++} gradient function.}

\item{const_init_ptr}{(\code{externalptr})\cr
External pointer to the \code{C++} constant initialization function.}
}
\description{
\code{LossCustomCpp} creates a custom loss by using
\code{Rcpp::XPtr} to set \code{C++} functions.
}
\section{Usage}{

\preformatted{
LossCustomCpp$new(loss_ptr, grad_ptr, const_init_ptr)
}
}

\examples{
\dontrun{
# Load loss functions:
Rcpp::sourceCpp(code = getCustomCppExample(example = "loss", silent = TRUE))

# Create new custom quadratic loss:
my_cpp_loss = LossCustomCpp$new(lossFunSetter(), gradFunSetter(), constInitFunSetter())
}
}
