% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{LossHuber}
\alias{LossHuber}
\title{Huber loss for regression tasks.}
\format{\code{\link{S4}} object.}
\description{
This loss can be used for regression with \eqn{y \in \mathrm{R}}.
}
\details{
\strong{Loss Function:}
\deqn{
  L(y, f(x)) = 0.5(y - f(x))^2 \ \ \mathrm{if} \ \ |y - f(x)| < d
}
\deqn{
  L(y, f(x)) = d|y - f(x)| - 0.5d^2 \ \ \mathrm{otherwise}
}
\strong{Gradient:}
\deqn{
  \frac{\delta}{\delta f(x)}\ L(y, f(x)) = f(x) - y \ \ \mathrm{if} \ \ |y - f(x)| < d
}
\deqn{
  \frac{\delta}{\delta f(x)}\ L(y, f(x)) = -d\mathrm{sign}(y - f(x)) \ \ \mathrm{otherwise}
}
}
\section{Usage}{

\preformatted{
LossHuber$new()
LossHuber$new(delta)
LossHuber$new(offset, delta)
}
}

\section{Arguments}{

\describe{
\item{\code{offset} [\code{numeric(1)}]}{
  Numerical value which can be used to set a custom offset. If so, this
  value is returned instead of the loss optimal initialization.
}
\item{\code{delta} [\code{numeric(1)}]}{
  Numerical value greater than 0 to specify the interval around 0 for the quadratic error measuring.
  Default is 1.
}
}
}

\examples{

# Create new loss object:
huber_loss = LossHuber$new()
huber_loss

}
