% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{LossQuantile}
\alias{LossQuantile}
\title{Quantile loss for regression tasks.}
\format{
\link{S4} object.
}
\arguments{
\item{offset}{(\code{numeric(1)} | \code{matrix()})\cr
Numerical value or matrix to set a custom offset. If used, this
value is returned instead of the loss optimal initialization.}

\item{quantile}{(\code{numeric(1)})\cr
Numerical value between 0 and 1 that defines the quantile that is modeled.}
}
\description{
This loss can be used for regression with \eqn{y \in \mathrm{R}}.
}
\details{
\strong{Loss Function:}
\deqn{
  L(y, f(x)) = h| y - f(x)|
}
\strong{Gradient:}
\deqn{
  \frac{\delta}{\delta f(x)}\ L(y, f(x)) = -h\mathrm{sign}( y - f(x))
}
\strong{Initialization:}
\deqn{
  \hat{f}^{[0]}(x) = \mathrm{arg~min}_{c\in R}\ \frac{1}{n}\sum\limits_{i=1}^n
  L(y^{(i)}, c) = \mathrm{quantile}(y, q)
}
}
\section{Usage}{

\preformatted{
LossAbsolute$new()
LossAbsolute$new(quantile)
LossAbsolute$new(offset, quantile)
}
}

\section{Inherited methods from Loss}{

\itemize{
\item \verb{$loss()}: \verb{matrix(), matrix() -> matrix()}
\item \verb{$gradient()}: \verb{matrix(), matrix() -> matrix()}
\item \verb{$constInit()}: \code{matrix() -> matrix()}
\item \verb{$calculatePseudoResiduals()}: \verb{matrix(), matrix() -> matrix()}
\item \verb{$getLossType()}: \verb{() -> character(1)}
}
}

\examples{

# Create new loss object:
quadratic_loss = LossQuadratic$new()
quadratic_loss

}
