% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{OptimizerAGBM}
\alias{OptimizerAGBM}
\title{Nesterov momentum}
\format{
\code{\link{S4}} object.
}
\description{
This class defines a new object which is used to conduct Nesterovs momentum as optimization technique.
}
\section{Usage}{

\preformatted{
OptimizerAGBM$new(momentum)
OptimizerAGBM$new(momentum, ncores)
}
}

\section{Arguments}{

\describe{
\item{\code{momentum} \link{\code{numeric(1)}}}{
Momentum term used to accelerate the fitting process. If chosen large, the algorithm trains
faster but also tends to overfit faster.
}
\item{\code{ncores} \link{\code{integer(1)}}}{
Number of cores used to fit the algorithm. Note that number of used cores
should be smaller or equal the number of base learner.
}
}
}

\examples{

optimizer = OptimizerAGBM$new(0.1)

}
