% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{OptimizerCoordinateDescent}
\alias{OptimizerCoordinateDescent}
\title{Coordinate Descent}
\format{\code{\link{S4}} object.}
\description{
This class defines a new object for the greedy optimizer. The optimizer
just calculates for each base-learner the sum of squared errors and returns
the base-learner with the smallest SSE.
}
\section{Usage}{

\preformatted{
OptimizerCoordinateDescent$new()
}
}

\examples{

# Define optimizer:
optimizer = OptimizerCoordinateDescent$new()

}
