% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{OptimizerCoordinateDescent}
\alias{OptimizerCoordinateDescent}
\title{Greedy Optimizer}
\format{\code{\link{S4}} object.}
\description{
This class defines a new object for the greedy optimizer. The optimizer
just calculates for each base-learner the sum of squared errors and returns
the base-learner with the smallest SSE.
}
\section{Usage}{

\preformatted{
OptimizerCoordinateDescent$new()
}
}

\section{Details}{


  This class is a wrapper around the pure \code{C++} implementation. To see
  the functionality of the \code{C++} class visit
  \url{https://schalkdaniel.github.io/compboost/cpp_man/html/classoptimizer_1_1_greedy_optimizer.html}.
}

\examples{

# Define optimizer:
optimizer = OptimizerCoordinateDescent$new()

}
