% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{OptimizerCoordinateDescentLineSearch}
\alias{OptimizerCoordinateDescentLineSearch}
\title{Coordinate Descent with line search}
\format{\code{\link{S4}} object.}
\description{
This class defines a new object which is used to conduct Coordinate Descent with line search.
The optimizer just calculates for each base-learner the sum of squared error and returns
the base-learner with the smallest SSE. In addition, this optimizer computes
a line search to find the optimal step size in each iteration.
}
\section{Usage}{

\preformatted{
OptimizerCoordinateDescentLineSearch$new()
}
}

\section{Details}{


  This class is a wrapper around the pure \code{C++} implementation. To see
  the functionality of the \code{C++} class visit
  \url{https://schalkdaniel.github.io/compboost/cpp_man/html/classoptimizer_1_1_greedy_optimizer.html}.
}

\examples{

# Define optimizer:
optimizer = OptimizerCoordinateDescentLineSearch$new()

}
