% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{OptimizerCosineAnnealing}
\alias{OptimizerCosineAnnealing}
\title{Coordinate descent with cosine annealing}
\format{
\link{S4} object.
}
\arguments{
\item{ncores}{(\code{integer(1)})\cr
Number of cores over which the base learner fitting is distributed.}

\item{nu_min}{(\code{numeric(1)})\cr
Minimal learning rate.}

\item{nu_max}{(\code{numeric(1)})\cr
Maximal learning rate.}

\item{cycles}{(\code{integer(1)})\cr
Number of annealing cycles form \code{nu_max} to \code{nu_min} between 1 and anneal_\code{anneal_iter_max}.
\verb{anneal_iter_max (}integer(1)\verb{)\\cr Maximal number of iterations for which the annealing is conducted. }nu_min\verb{is used as fixed learning rate after}anneal_iter_max`.}
}
\description{
Same as \link{OptimizerCoordinateDescent} but with a cosine annealing scheduler to
adjust the learning rate during the fitting process.
}
\section{Usage}{

\preformatted{
OptimizerCosineAnnealing$new()
OptimizerCosineAnnealing$new(ncores)
OptimizerCosineAnnealing$new(nu_min, nu_max, cycles, anneal_iter_max, cycles)
OptimizerCosineAnnealing$new(nu_min, nu_max, cycles, anneal_iter_max, cycles, ncores)
}
}

\section{Fields}{

This class doesn't contain public fields.
}

\section{Methods}{

\itemize{
\item \verb{$getOptimizerType()}: \verb{() -> character(1)}
\item \verb{$getStepSize()}: \verb{() -> numeric()}
}
}

\examples{

# Define optimizer:
optimizer = OptimizerCosineAnnealing$new()

}
