% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ResponseBinaryClassif}
\alias{ResponseBinaryClassif}
\title{Create response object for binary classification.}
\format{\code{\link{S4}} object.}
\description{
\code{ResponseBinaryClassif} creates a response object that are used as target during the 
fitting process.
}
\section{Usage}{

\preformatted{
ResponseBinaryClassif$new(target_name, response)
ResponseBinaryClassif$new(target_name, response, weights)
}
}

